feat: implement pagination in Products component
Introduce pagination functionality in the Products component to enhance the user experience. The following changes were made:

- Imported Pagination components from @/components/ui/pagination.
- Utilized useSearchParams from next/navigation to dynamically update the page parameter in the URL.
- Created a Pagination component with Previous and Next links, allowing users to navigate between different pages.
- Adjusted the rendering logic in the Products component to display a subset of products based on the active page.
- Improved styling for a responsive grid layout to showcase the paginated product cards.

This commit addresses the need for a more user-friendly product listing by introducing pagination, making it easier for users to explore a large number of products.
chore: update dependencies to latest versions

Update dependencies to the latest versions:
- @trpc/react-query: from 10.45.0 to 11.0.0-next-beta.260
- @tanstack/react-query: from 4.18.0 to 5.18.1
- @trpc/client: from 10.45.0 to 11.0.0-next-beta.260
- @trpc/server: from 10.45.0 to 11.0.0-next-beta.260

Peer dependencies have also been updated:
- @tanstack/react-query: ^4.18.0 → ^5.18.1

Additionally, the resolution field has been updated with the latest integrity hashes.

This commit reflects the need to stay up-to-date with the latest versions and peer dependencies, ensuring compatibility and taking advantage of potential improvements and bug fixes.